+++ lib/dropdown/drop_down_view_wrapper.dart
@@ -3,14 +3,19 @@ import 'package:flutter/material.dart';
 import 'drop_down_typedef.dart';
 
 class DropDownViewWrapper extends DropDownViewStatelessWidget {
+  final double width;
   final double height;
   final Widget child;
 
   const DropDownViewWrapper({
     super.key,
+    required this.width,
     required this.height,
     required this.child,
   });
+
+  @override
+  double get actualWidth => width;
 
   @override
   double get actualHeight => height;